+++ src/fixes/index.js
@@ -1,0 +1,5 @@
+import customSelectorsMissingColon from "./custom-selectors-missing-colon"
+
+export default {
+  customSelectorsMissingColon,
+}