+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-
+- Added: warning to deprecate previous (wrong) custom selectors syntax.
+Previously they were working with **and without** pseudo syntax ':'
+[which is incorrect according to specs](https://github.com/postcss/postcss-custom-selectors/issues/5#issuecomment-90774654)
+Now you must use `@custom-selector :--{name}` syntax instead of `@custom-selector --{name}`
+The support of syntax without : and the warning message will be remove in the next major release
+([#97](https://github.com/cssnext/cssnext/issues/97))
 - Added: `plugins` option that allows you to pipe your own transformations ([#118](https://github.com/cssnext/cssnext/issues/118))
 - Added: `messages` option that allows you to see messages of transformations ([#88](https://github.com/cssnext/cssnext/issues/88))